Change versions back to 1.3.0 and (4.0.3).
This PR has two SOFA updates: 20160503 and 20160503_a. I had
set 1.3 (4.0.3) to correspond to 20160503 and 1.4 (5.0.4) to
20160503_a. There is no need for this as liberfa is not tied
to release schedule of SOFA. I am setting this release of
liberfa to 1.3 (4.0.3) (i.e., increment from 1.2 (3.0.2),
and it corresponds to SOFA 20160503_a.

In README.rst, I made the note that ERFA 1.3 has no changes with
respect to SOFA 20160503_a.