Make day + fraction separation more robust
jd2cal.c assumed that for any x, x-fmod(x, 1.0) returns a value not smaller
than x. Due to rounding effects, this does not always need to be the case.
Especially on i386, the result might be slightly smaller. Therefore, the code
floor(x-fmod(x, 1.0)) is not guaranteed to return the same (expected) value
on all platforms.

Since we know, that x-fmod(x,1.0) is close to an int, we just replace this
by a normal rounding. This is done by adding 0.5, to stay away from such
modern stuff like rint, which hopefully makes the code acceptable to
old-style programmers from SOFA as well ;-)