Improve accuracy of eraStarpv()
In the relativistic correction, the term

 del = sqrt(1+x) - 1

is used with a rather small value of x (often 1e-8 or such). This gives
inaccurate results because of the subtraction of 1 from a number very close
to 1.

To improve this, the right term is extended by sqrt(1+x)+1, which gives

 del = x / (sqrt(1+x) + 1)

without a subtraction of two equally large numbers and therefore a better
accuracy. This helps on i386 FPU.

The test is adjusted as well.