Preparing for release 2.0.0
With the change to how include files are dealt with, effectively macros
have changed location and people will need to adjust their code, so a
new major version seems warranted.  On the other hand, for the actual
library, only new functions were added, so the library is backwards
compatible and hence current and age are increased.